refactor: fetch의 책임 분리
