+++ src/app.feature/userManagement/module/useFetchUser.ts
@@ -1,0 +1,38 @@
+import { BASE_URL } from "@/app.module/api/environment";
+import axios from "axios";
+import { useEffect, useState } from "react";
+
+type User = {
+  id: number;
+  name: string;
+  email: string;
+  role: 'admin' | 'user';
+  lastLogin: string;
+};
+
+const useFetchUser = () => {
+  const [users, setUsers] = useState<User[]>([]);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
+
+  const fetchUsers = async () => {
+    try {
+      setLoading(true);
+      const response = await axios.get<User[]>(`${BASE_URL}/users`);
+      setUsers(response.data);
+      setLoading(false);
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    } catch (err) {
+      setError('Failed to fetch users');
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchUsers();
+  }, [])
+
+  return { users, loading, error, refetch: fetchUsers };
+}
+
+export default useFetchUser;
